+++ src/base/b-dummy/b-dummy.ts
@@ -12,8 +12,9 @@ import { ResizeWatcher } from 'core/dom/resize-observer';
 
 import iLockPageScroll from 'traits/i-lock-page-scroll/i-lock-page-scroll';
 
-import iData, { component } from 'super/i-data/i-data';
+import iData, { component, wait, hook } from 'super/i-data/i-data';
 import { Directives, Modules } from 'base/b-dummy/interface';
+import iObserveDOM from 'traits/i-observe-dom/i-observe-dom';
 
 const
 	inViewMutation = inViewFactory('mutation'),
@@ -29,7 +30,7 @@ export * from 'base/b-dummy/interface';
 	}
 })
 
-export default class bDummy extends iData implements iLockPageScroll {
+export default class bDummy extends iData implements iLockPageScroll, iObserveDOM {
 	/** @see [[iLockPageScroll.lock]] */
 	lock(): Promise<void> {
 		return iLockPageScroll.lock(this);
@@ -38,6 +39,22 @@ export default class bDummy extends iData implements iLockPageScroll {
 	/** @see [[iLockPageScroll.unlock]] */
 	unlock(): Promise<void> {
 		return iLockPageScroll.unlock(this);
+	}
+
+	/** @see [[iObserveDOM.prototype.initDOMObservers]] */
+	@hook('mounted')
+	@wait('ready')
+	initDOMObservers(): void {
+		iObserveDOM.observe(this, {
+			node: this.$el!,
+			childList: true,
+			subtree: true
+		});
+	}
+
+	/** @see [[iObserveDOM.prototype.emitDOMChange]] */
+	onDOMChange(): void {
+		iObserveDOM.emitDOMChange(this);
 	}
 
 	/**
@@ -57,7 +74,8 @@ export default class bDummy extends iData implements iLockPageScroll {
 	 */
 	get modules(): Modules {
 		return {
-			resizeWatcher: ResizeWatcher
+			resizeWatcher: ResizeWatcher,
+			iObserveDOM
 		};
 	}
 }